+++ src/LevelOne.h
@@ -20,9 +20,9 @@ public:
 		// Timer
 	std::unordered_map<Uint32, TimerObject*> player_timers;
 	TimerObject* spawn_pu_timer = nullptr;
-	void onPlayerTimerFinish(Uint8 flag);
-	void onEndGameTimerFininsh(Uint8 flag);
-	void onSpawnPowerUpTimerFinish(Uint8 flag);
+	void onPlayerTimerFinish(Uint32 flag);
+	void onEndGameTimerFininsh(Uint32 flag);
+	void onSpawnPowerUpTimerFinish(Uint32 flag);
 
     // GameObjects
 	Tower* tower = nullptr;
